better rounding for teleport coords

+++ src/Command/CmdBack.cs
@@ -7,7 +7,7 @@ namespace BasicCommands.Command;
 
 public class CmdBack : AbstractCommand {
     public override TextCommandResult Execute(BasicPlayer sender, TextCommandCallingArgs args) {
-        Vec3i lastPos = sender.LastPos;
+        Vec3d lastPos = sender.LastPos;
         if (lastPos == null) {
             return TextCommandResult.Success(Lang.Get("back-empty"));
         }
